AddedSession

+++ WebProject/WebProject/Controllers/AdminController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Web;
 using System.Web.Mvc;
 using WebProject.BusinessLogic.MainBL;
+using WebProject.Domain.Enum;
 using WebProject.ModelAccessLayer.Model;
 
 
@@ -14,28 +15,36 @@ namespace WebProject.Controllers
         // GET: AdminAddProduct
         public ActionResult NewProduct()
         {
-            return View();
+            if (Session["UserData"] != null && ((UserData)Session["UserData"]).StatusUser == StatusUser.Admin)
+                return View();
+            return RedirectToAction("Index", "Home");
         }
+
         public ActionResult ViewProducts()
         {
-            UserData userData = new UserData();
-            userData.ProductsAdmin = AccountController.admin.GetAllProducts();
-
-            return View(userData);
+            if (Session["UserData"] != null && ((UserData)Session["UserData"]).StatusUser == StatusUser.Admin)
+                return View((UserData)Session["UserData"]);
+            return RedirectToAction("Index", "Home");
         }
+
         public ActionResult ViewDelivery()
         {
-            UserData userData = new UserData();
-            userData.ProductsAdmin = AccountController.admin.GetAllProducts();
-
-            return View(userData);
+            if (Session["UserData"] != null && ((UserData)Session["UserData"]).StatusUser == StatusUser.Admin)
+                return View((UserData)Session["UserData"]);
+            return RedirectToAction("Index", "Home");
         }
-        
+
+
+        //в бизнес ложик обмен данными
         [HttpPost]
         public ActionResult NewProduct(Product product)
         {
+            if (Session["UserData"] == null || ((UserData)Session["UserData"]).StatusUser != StatusUser.Admin)
+                return RedirectToAction("Index", "Home");
+
             if (ModelState.IsValid)
             {
+                Console.WriteLine("All Good");
                 AccountController.admin.AddProduct(product);
 
                 return RedirectToAction("Index", "Home");
@@ -43,18 +52,24 @@ namespace WebProject.Controllers
             else
                 return View(product);
         }
-        
+
         [HttpPost]
         public ActionResult EditProduct(int id)
         {
-            Product editProduct =  AccountController.guest.GetProductById(id);
+            if (Session["UserData"] == null || ((UserData)Session["UserData"]).StatusUser != StatusUser.Admin)
+                return RedirectToAction("Index", "Home");
+
+            Product editProduct = AccountController.guest.GetProductById(id);
 
             return View(editProduct);
         }
 
         [HttpPost]
-        public ActionResult ReplaceProduct(Product product) 
+        public ActionResult ReplaceProduct(Product product)
         {
+            if (Session["UserData"] == null || ((UserData)Session["UserData"]).StatusUser != StatusUser.Admin)
+                return RedirectToAction("Index", "Home");
+
             if (ModelState.IsValid)
             {
                 AccountController.admin.EditProduct(product);
@@ -62,21 +77,27 @@ namespace WebProject.Controllers
                 return RedirectToAction("ViewProducts", "Admin");
             }
             else
-                return RedirectToAction("EditProduct","Admin",product.Id);
+                return RedirectToAction("EditProduct", "Admin", product.Id);
         }
 
         [HttpPost]
         public ActionResult DeleteProduct(int id)
         {
+            if (Session["UserData"] == null || ((UserData)Session["UserData"]).StatusUser != StatusUser.Admin)
+                return RedirectToAction("Index", "Home");
+
             AccountController.admin.DeleteProduct(id);
             return RedirectToAction("ViewProducts", "Admin");
         }
-        
+
         [HttpPost]
         public ActionResult DeleteDelivery(int id)
         {
-            AccountController.admin.DeleteOrderModel(id);   
-            return RedirectToAction("ViewDelivery","Admin");
+            if (Session["UserData"] == null || ((UserData)Session["UserData"]).StatusUser != StatusUser.Admin)
+                return RedirectToAction("Index", "Home");
+
+            AccountController.admin.DeleteOrderModel(id);
+            return RedirectToAction("ViewDelivery", "Admin");
         }
     }
 }
